requirements.txt
amqp==2.5.2
asgiref==3.2.7
beat==0.3.0
billiard==3.6.3.0
celery==4.4.2
dj-database-url==0.5.0
Django==3.0.6
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
flower==0.9.4
gunicorn==20.0.4
humanize==0.5.1
importlib-metadata==1.6.0
kombu==4.6.8
Pillow==7.1.2
python-http-client==3.2.7
pytz==2020.1
redis==3.5.2
sendgrid==6.3.1
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
psycopg2==2.7.2
